+++ app/models/specialty_category.rb
@@ -1,0 +1,4 @@
+class SpecialtyCategory < ApplicationRecord
+  has_many :specialty_to_categories, dependent: :destroy
+  has_many :specialties, through: :specialty_to_categories
+end