+++ app/models/specialty_to_category.rb
@@ -1,0 +1,4 @@
+class SpecialtyToCategory < ApplicationRecord
+  belongs_to :specialty
+  belongs_to :specialty_category
+end